refactor: Loose coupling,folder modules and classes
